Refactored to int64

+++ MiNTT8/mintt8.go
@@ -10,37 +10,37 @@ func MinNTT8(input [ndiv8 * m]byte) [864]byte {
 
 }
 
-func ntt_sum(input [ndiv8 * m]byte) [N]int {
+func ntt_sum(input [ndiv8 * m]byte) [N]int64 {
 
-	var solution [N]int
-	for i := 0; i < m; i++ {
+	var solution [N]int64
+	for i := int64(0); i < m; i++ {
 		x := NTT8_TABLE[input[i]]
-		for j := 0; j < d; j++ {
-			for k := 0; k < n; k++ {
+		for j := int64(0); j < d; j++ {
+			for k := int64(0); k < n; k++ {
 				solution[n*j+k] = solution[n*j+k] + x[k]*A[i][n*j+k]
 			}
 
 		}
 	}
-	for i := 0; i < N; i++ {
+	for i := int64(0); i < N; i++ {
 		solution[i] = util.Mod(solution[i], q)
 	}
 
 	return solution
 }
 
-//Assume vaules haveaady been ModQ'd
-func ChangeBase(val [N]int) [864]byte {
+// Assume vaules haveaady been ModQ'd
+func ChangeBase(val [N]int64) [864]byte {
 
 	var output [864]byte
 
-	for i := 0; i < N; i++ {
+	for i := int64(0); i < N; i++ {
 		output[i] = byte(val[i])
 		val[i] = val[i] >> 8
 	}
 
-	for i := 0; i < Ndiv8; i++ {
-		for k := 0; k < 8; k++ {
+	for i := int64(0); i < Ndiv8; i++ {
+		for k := int64(0); k < 8; k++ {
 			output[N+i] = output[N+i] | byte(val[8*i+k]>>k)
 		}
 	}
